added DAL pre-controller to user domain
