+++ src/api/apiDAL.ts
@@ -1,0 +1,7 @@
+export abstract class ApiReqHandler<ControllerType> {
+    constructor(private _controller: ControllerType) {}
+
+    public get controller() {
+        return this._controller;
+    }
+}
