+++ src/api/user/index.ts
@@ -1,6 +1,7 @@
 import initRoutes from './routes';
 import { Server } from '../../types/server';
 import UserController from './controller';
+import UserDal from './userDAL';
 import { ApiEnterOptions } from '../../types/common';
 
 export default async function init(
@@ -11,6 +12,7 @@ export default async function init(
     }: ApiEnterOptions
 ): Promise<boolean> {
     const userController = new UserController(userModel, serverConfigs);
+    const userDal = new UserDal(userController);
     server.bind(userController);
     await initRoutes(server, serverConfigs, userController);
